changed: vector values from doubles to ints

+++ libs/vector.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include <math.h>
 
-double vec_length(double vec[]){
+double vec_length(int vec[]){
   double s = 0;
   for(int i = 0; i < 3; i++){
     s += pow(vec[i], 2);
@@ -12,11 +12,13 @@ double vec_length(double vec[]){
   return sqrt(s);
 }
 
-void vec_add(double vec[], double vec1[], double v[]){
+void vec_add(int vec[], int vec1[], int v[]){
   for(int i = 0; i < 3; i++){
     v[i] = vec[i] + vec1[i];
   }
   return;
 }
+
+
 
 #endif
